# Make printing errors faster