### What does this PR do?

This makes printing errors faster for transpiler errors and for runtime errors

`perf trace` after:

```zig
1 | throw new Error("This is an error");
          ^
error: This is an error
      at /tmp/runtime-error.ts:1:7
      at asyncFunctionResume (:1:21)
      at promiseReactionJobWithoutPromiseUnwrapAsyncContext (:1:21)
      at promiseReactionJob (:1:21)
   148.938 ( 0.005 ms): bun-debug/3886354 write(fd: 2</dev/pts/9>, buf: 0x7ffd5a0a827c, count: 686)             = 686
```

`pef trace` before:

```zig
   148.056 1( 0.005 ms | ): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362eb1ce1, count: 8)               = 8
   148.063 ( 0.001 ms): bun-debug/3884968 throwwrite(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef9f7, count: 1)               = 1
    148.065 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362eb1cec, count: 7)               = 7new
   148.069 (  0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d96d41, count: 4Error)               = 4
   148.073 (( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x55636d3cff53, count: 5)               = 5
"This is an error"   148.076 ( 0.002 ms): bun-debug/)3884968 write(fd: 2</dev/pts/9>, buf: 0x20000440010, count: 5)                = ;5

   148.078 ( 0.001 ms          ): bun-debug/^
3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d96d4b, count: 4)               = 4
   148.080 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x20000440015, count: 1)                = 1
error   148.083 (:  0.002 ms): bun-debug/This is an error3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d96d41, count: 4
)               = 4
   148.085 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x55636d3cff53, count: 5)               = 5
   148.087 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x20000440016, count: 3)                = 3
   148.089 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d96d4b, count: 4)               = 4
   148.091 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x20000440019, count: 1)                = 1
   148.093 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d96dc1, count: 8)               = 8
   148.095 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x2000044001a, count: 5)                = 5
   148.097 ( 0.002 ms): bun-debug/      at 3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d96dcc, count: 4)               = 4
/tmp/runtime-error.ts   148.100 ( 0.002 ms): bun-debug/:3884968 write(fd: 2</dev/pts/9>, buf: 0x2000044001f, count: 1)                = 11
:   148.102 ( 0.002 ms7): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d97129, count: 9)               = 9

   148.105 ( 0.001 ms      at ): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x20000440020, count: 18)               = 18asyncFunctionResume
   148.107 ( ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362d97135, count: 4)               = 4
:   148.109 (1 0.002 ms): :bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x20000440032, count: 121)                = 1
   148.112 ()
 0.002 ms): bun-debug/3884968       at write(fd: 2</dev/pts/9>, buf: 0x556362d96f59, count: 13)              = 13
promiseReactionJobWithoutPromiseUnwrapAsyncContext   148.114 ( 0.001 ms): bun-debug/ (3884968 write(fd: 2</dev/pts/9>, buf: 0x556362eb1cf5, count: 1)               = 1
:   148.116 ( 0.002 ms1): bun-debug/:3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34efa18, count: 10)              = 2110
   148.118 ()
 0.001 ms): bun-debug/      at 3884968 write(fd: 2</dev/pts/9>, buf: 0x556362eb1de1, count: 15)              = 15
promiseReactionJob   148.121 ( 0.002 ms): bun-debug/3884968  (write(fd: 2</dev/pts/9>, buf: 0x556362eb1941, count: 9)               = 9
   148.129 (: 0.001 ms): 1bun-debug/3884968 :write(fd: 2</dev/pts/9>, buf: 0x55636d3c6fcf, count: 5)               = 215
   148.131 )
( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362eb194c, count: 18)              = 18
   148.134 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7f1025174314, count: 16)              = 16
   148.136 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362eb1961, count: 5)               = 5
   148.178 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acffa1, count: 21)              = 21
   148.181 ( 0.004 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362e5ffb6, count: 9)               = 9
   148.185 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7f10250d6e64, count: 21)              = 21
   148.187 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x5563629862b9, count: 4)               = 4
   148.189 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def669, count: 14)              = 14
   148.191 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2c7, count: 1)               = 1
   148.193 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def67a, count: 14)              = 14
   148.196 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2c7, count: 1)               = 1
   148.198 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def68b, count: 4)               = 4
   148.200 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acffb8, count: 1)               = 1
   148.204 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe59, count: 21)              = 21
   148.206 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfb79, count: 12)              = 12
   148.209 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x5563630c0f74, count: 19)              = 19
   148.211 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfb87, count: 4)               = 4
   148.213 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe70, count: 10)              = 10
   148.216 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362e5ffb6, count: 9)               = 9
   148.218 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x5563629862b9, count: 4)               = 4
   148.220 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def669, count: 14)              = 14
   148.222 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2b7, count: 1)               = 1
   148.223 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def67a, count: 14)              = 14
   148.226 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2b6, count: 2)               = 2
   148.228 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def68b, count: 4)               = 4
   148.230 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe7c, count: 10)              = 10
   148.232 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe59, count: 21)              = 21
   148.234 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfb79, count: 12)              = 12
   148.236 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556363689bc5, count: 50)              = 50
   148.239 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfb87, count: 4)               = 4
   148.240 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe70, count: 10)              = 10
   148.243 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362e5ffb6, count: 9)               = 9
   148.245 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x5563629862b9, count: 4)               = 4
   148.247 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def669, count: 14)              = 14
   148.249 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2b7, count: 1)               = 1
   148.251 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def67a, count: 14)              = 14
   148.252 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2b6, count: 2)               = 2
   148.255 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def68b, count: 4)               = 4
   148.257 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe7c, count: 10)              = 10
   148.260 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe59, count: 21)              = 21
   148.263 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfb79, count: 12)              = 12
   148.265 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x55636327b42e, count: 18)              = 18
   148.267 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfb87, count: 4)               = 4
   148.269 ( 0.002 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe70, count: 10)              = 10
   148.271 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362e5ffb6, count: 9)               = 9
   148.273 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x5563629862b9, count: 4)               = 4
   148.275 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def669, count: 14)              = 14
   148.277 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2b7, count: 1)               = 1
   148.278 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def67a, count: 14)              = 14
   148.280 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x7ffcd34ef2b6, count: 2)               = 2
   148.282 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362def68b, count: 4)               = 4
   148.284 ( 0.001 ms): bun-debug/3884968 write(fd: 2</dev/pts/9>, buf: 0x556362acfe7c, count: 10)              = 10

```



### How did you verify your code works?

CI